move mock/utils file
